requirements.txt
numpy>=1.26
scipy>=1.12
PyYAML>=6.0
pyserial>=3.5
pyserial-asyncio>=0.6
